First code !
Index.js is the heart of the (aspiring) library. We have a parent Class Shape, which is extended by all of our shapes; for now it is just ball and rectangle.
The html file is there just for me to test and show you how to set up the library.

It is pretty simple, once you have declared canvas and his context, you just have to declare instances of our classes.
Until I find a better way to do it, you'll have to add manually all the instances into an array with .concat(). Good thing we can concat multiple array at once.

Once everything is declared, you draw the shapes, make the ones you want move and... That's it for now.

I'll be gradually adding collisions, new shapes but the more complex the shape, the harder it'll be to detect collision.