Added canvas boundaries
This is the first step for collisions. Setting boundaries.
For now it is not configurable, but in the end I'd like to let the user choose which side of the canvas they want to set boundaries.

To do so, you just have to specify if you want a shape to be affected by the boundaries. It is the last property of the constructor.

Finnaly each shape will have its own setBoundaries method beacause of the differences of collision detection.