Added Collision detection
I added the collision dectection for the circles. (circle-circle and circle-rectangle)
When the collision is detected, the function returns true.

To add a collision, you just have to call the method .addCollision on an object, giving as parameter an array with all the objects you want to collide with.

I think I am going to simplify the creation of objects and add some methods to define some property (as i did to add colliding objects)
Though, I am happy with the way things are going.