Added methods to clarify instanciation of shapes and a callback for collision detection
Removed velocities and boundaries from contructor method  and created separeted methods to declare these values.

I added a method to regroup all the instances in an array, I don't like the way it is done and think there's probably a way to do it automatically on instanciation of the class. But for now, it is enough.

I also used arguments to get any numbers of arguments in some methods

I changed the name of Ball class (it is Circle now).