Added Boundaries and rectangle to rectangle collision detection
Added these two algorythms, but found that there is a problem if in our script there is more than one callback function.

The problem is that the last callback function overide all of the others. In index.html you can see line 56 and line 61, I have two different callback function. Since the last one makes the shape blue, all collision detected will make the shape blue; even though it should have become pink.

I don't really know where this problem comes from, if it's because these are two anonymous function (I tried naming one of them, and the problem persisted). I think it comes from our script and not the algorythm because in index.js line 306, I added a console log that shows only on the collision between rect and square. This console log works as expected, only when these two shapes collide.
So the callback function is not called on line 304 but still the console log line 62 of index.html is called.
I will have to spend some time fixing this.