Fixed the callbacks of collision detection, added rect/circle collision detection
I fixed the problem I had with the callback functions in the HTML. Problem was each and every callback functions were activated even though it shouldn't.
I thouhght the probleme came from the html but it wasn't, it came from the moveShape method.
in index.js line 58 and 61, I used to call the method setCollision on each and every shapes created which caused to call undesired callback functions activation.
I changed it to just call setCollision on the shape that called moveShape. So I just changed instances[i] to this.
It was kind of a careless mistake, but I'm glad I find it.

I also finished the setCollision method for rectangle which now detect collisions between a rectangle and a cirlce. It wasn't hard at all, I just copy the circle one and swap this and shape.

So I think that's it for v1.0.0 I think